web/requirements.txt
altair
astroid
attrs
blinker
cachetools
certifi
charset-normalizer
click
commonmark
decorator
dill
entrypoints
filelock
gitdb
GitPython
huggingface-hub
idna
ImageHash
importlib-metadata
isort
Jinja2
jsonschema
lazy-object-proxy
MarkupSafe
mccabe
numpy
packaging
pandas
Pillow
platformdirs
protobuf
pyarrow
pydeck
Pygments
pylint
Pympler
pyparsing
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
PyWavelets
PyYAML
regex
requests
rich
scipy
semver
six
smmap
streamlit
streamlit-ext
streamlit-option-menu
streamlit-vis-timeline
tokenizers
toml
tomli
tomlkit
toolz
tornado
tqdm
transformers
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
wrapt
zipp
